+++ api/src/routes/UserRouter.js
@@ -1,5 +1,6 @@
 const { Router } = require("express");
-
+const multer = require ('multer');
+const path = require('path');
 
 const {
   getUserHandler,
@@ -15,27 +16,33 @@ const {
 
 const UserRouter = Router();
 
-// const checkUserProperties = (req, res, next) => {
-//   const { email, password, username } = req.body;
+const storage = multer.diskStorage({
+  destination: path.join(__dirname, "assets"),
+  filename: (req, file, cb) => {
+      cb(null, Date.now() + path.extname(file.originalname));
+  }
+});
 
-//   if (email && password && username) {
-//     return postUserHandler(req, res, next);
-//   }
-
-//   if (email && password) {
-//     return getUserHandler(req, res, next);
-//   }
-
-//   return res.status(400).json({ error: "Falta información en la solicitud" });
-// };
+const uploadImage = multer({
+  storage: storage,
+  fileFilter: (req, file, cb) => {
+      const filetypes = /jpeg|jpg|png|gif/;
+      const mimetype = filetypes.test(file.mimetype);
+      const extname = filetypes.test(path.extname(file.originalname));
+      if(mimetype && extname) {
+          return cb(null, true);
+      }
+      cb("Error: El archivo no es de tipo imagen.");
+  },
+}).single("image");
 
 // UserRouter.use(checkUserProperties);
 UserRouter.get("/", getAllUsersHandler);
 UserRouter.get("/:user_id", getUserById);
 UserRouter.post("/", postUserHandler)
 // UserRouter.get("/:id_user", getAllUsersHandler);
-UserRouter.put("/edit", putEditUserHandler)
-UserRouter.delete("/status/:id_user",putStatusUserHandler)
+UserRouter.put("/edit", uploadImage, putEditUserHandler)
+UserRouter.delete("/status/:id_user", putStatusUserHandler)
 UserRouter.post("/restore/:id_user", restoreStatusUserHandler)
 UserRouter.put("/", putRolUserHandler);
 UserRouter.get('/email/:email', getUserHandler)
